Interface RouterEvents added
In the old design there was a SceneManager and he also called OnEnter / OnLeave
on the affected scenes. This is a very useful feature but got lost with the last
refactoring. The replacement of the very specific SceneManager is the more
general Router and with this commit the router got a "hidden" feature to fill
this gap. How does ist work? The router simply checks if the previous / current
object implements the RouterEvents interface and calls OnEnter / OnLeave only if
it's possible.