Building block to support different android markets
I've added a new preprocessor constant, #BONO_ANDROID_MARKET, that can be used
to detect if the app was published via google or amazon. There are also two new
"normal" Monkey constants in the Target module. The "default publisher" is
google but here are two short examples:

    Import bono
    If Target.IS_MARKET_GOOGLE Then Print "True because default"

--

    Import bono
    #BONO_ANDROID_MARKET="amazon"
    If Not Target.IS_MARKET_GOOGLE Then Print "True because we've changed it"

In future commits this handling can be ...
- ... used to detect the right IAP backend
- ... used to generate the right market URI
- ... extended to support even more android markets

One might wonder why this is handled with a global preprocessor flag and the
answer is simple: Because this is the only reliable way to do so. And yes. I've
used google and read through a lot of comments on stackoverflow.